requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
dj-static==0.0.6
Django==3.2.3
django-bootstrap4==3.0.1
django-stdimage==5.3.0
gunicorn==20.1.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
